Dropping pure hibernate spec and migration to jpa

+++ nginx-admin-ui/src/main/java/com/jslsolucoes/nginx/admin/repository/AccessLogRepository.java
@@ -25,5 +25,5 @@ public interface AccessLogRepository {
 	
 	public List<AccessLog> listAll(Integer firstResult,Integer maxResults);
 
-	public Integer count();
+	public Long count();
 }
